[fix] Separate the test modules
